pyproject.toml
[tool.poetry]
name = "sajda-feed"
version = "0.1.0"
description = ""
authors = ["Beibarys Moldbekov <biba4756@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
attrs = "^23.2.0"
pillow = "^10.2.0"
coreapi = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
